Refactor state selectors into focused modules
