+++ tests/test_state_selectors_ui.py
@@ -1,0 +1,103 @@
+from tests import state_selectors_cases as cases
+
+test_select_attribute_dialog_state_formats_selected_entry = (
+    cases.test_select_attribute_dialog_state_formats_selected_entry
+)
+test_select_config_dialog_state_formats_editor_lines = (
+    cases.test_select_config_dialog_state_formats_editor_lines
+)
+test_select_config_dialog_state_shows_custom_editor_command_hint = (
+    cases.test_select_config_dialog_state_shows_custom_editor_command_hint
+)
+test_select_conflict_dialog_state_formats_create_directory_conflict = (
+    cases.test_select_conflict_dialog_state_formats_create_directory_conflict
+)
+test_select_conflict_dialog_state_formats_delete_confirmation = (
+    cases.test_select_conflict_dialog_state_formats_delete_confirmation
+)
+test_select_conflict_dialog_state_formats_extract_confirmation = (
+    cases.test_select_conflict_dialog_state_formats_extract_confirmation
+)
+test_select_conflict_dialog_state_formats_first_conflict = (
+    cases.test_select_conflict_dialog_state_formats_first_conflict
+)
+test_select_conflict_dialog_state_formats_permanent_delete_confirmation = (
+    cases.test_select_conflict_dialog_state_formats_permanent_delete_confirmation
+)
+test_select_conflict_dialog_state_formats_rename_conflict = (
+    cases.test_select_conflict_dialog_state_formats_rename_conflict
+)
+test_select_conflict_dialog_state_formats_zip_confirmation = (
+    cases.test_select_conflict_dialog_state_formats_zip_confirmation
+)
+test_select_help_bar_defaults_to_browsing_shortcuts = (
+    cases.test_select_help_bar_defaults_to_browsing_shortcuts
+)
+test_select_help_bar_for_attribute_dialog = (
+    cases.test_select_help_bar_for_attribute_dialog
+)
+test_select_help_bar_for_busy_mode = cases.test_select_help_bar_for_busy_mode
+test_select_help_bar_for_delete_confirmation = (
+    cases.test_select_help_bar_for_delete_confirmation
+)
+test_select_help_bar_for_name_conflict = cases.test_select_help_bar_for_name_conflict
+test_select_help_bar_for_paste_conflict_uses_generic_guidance = (
+    cases.test_select_help_bar_for_paste_conflict_uses_generic_guidance
+)
+test_select_help_bar_for_permanent_delete_confirmation = (
+    cases.test_select_help_bar_for_permanent_delete_confirmation
+)
+test_select_help_bar_for_split_terminal_focus = (
+    cases.test_select_help_bar_for_split_terminal_focus
+)
+test_select_help_bar_state_for_bookmarks_palette = (
+    cases.test_select_help_bar_state_for_bookmarks_palette
+)
+test_select_help_bar_state_for_command_palette = (
+    cases.test_select_help_bar_state_for_command_palette
+)
+test_select_help_bar_state_for_config_editor = (
+    cases.test_select_help_bar_state_for_config_editor
+)
+test_select_help_bar_state_for_file_search_palette = (
+    cases.test_select_help_bar_state_for_file_search_palette
+)
+test_select_help_bar_state_for_filter_mode = (
+    cases.test_select_help_bar_state_for_filter_mode
+)
+test_select_help_bar_state_for_go_to_path_palette_mentions_tab_completion = (
+    cases.test_select_help_bar_state_for_go_to_path_palette_mentions_tab_completion
+)
+test_select_help_bar_state_for_grep_search_palette = (
+    cases.test_select_help_bar_state_for_grep_search_palette
+)
+test_select_help_bar_state_for_history_palette = (
+    cases.test_select_help_bar_state_for_history_palette
+)
+test_select_input_bar_state_for_create_mode = (
+    cases.test_select_input_bar_state_for_create_mode
+)
+test_select_input_bar_state_for_filter_mode = (
+    cases.test_select_input_bar_state_for_filter_mode
+)
+test_select_input_bar_state_for_pending_key_sequence = (
+    cases.test_select_input_bar_state_for_pending_key_sequence
+)
+test_select_input_bar_state_formats_extract_mode = (
+    cases.test_select_input_bar_state_formats_extract_mode
+)
+test_select_input_bar_state_formats_zip_mode = (
+    cases.test_select_input_bar_state_formats_zip_mode
+)
+test_select_input_bar_state_keeps_active_filter_visible_after_confirm = (
+    cases.test_select_input_bar_state_keeps_active_filter_visible_after_confirm
+)
+test_select_split_terminal_state_builds_terminal_view = (
+    cases.test_select_split_terminal_state_builds_terminal_view
+)
+test_select_status_bar_exposes_notification_level = (
+    cases.test_select_status_bar_exposes_notification_level
+)
+test_select_status_bar_shows_split_terminal_focus_when_idle = (
+    cases.test_select_status_bar_shows_split_terminal_focus_when_idle
+)
